+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## DSL Improvements
+- Added support for calling python `Handler` methods via `on_enter` and `on_enter_once` hooks.
 - Fixed formatting of the "No timeout configured" warning
-